requirements.txt
Django==2.2
django-redis-cache==2.1.0
celery==4.3.0
httplib2==0.9.2
requests==2.18.0
packaging==19.1
Pillow==4.0.0
mysqlclient==1.4.4
django-mysql==3.2.0
tzlocal==1.3
six==1.10.0
django-environ==0.4.1
python-json-logger==0.1.8
tornado==4.2
gunicorn==19.6.0
sentry-sdk==0.10.2
passlib==1.7.1
bcrypt==3.1.7
django-prettyjson==0.4.1
standardjson==0.3.1
django-jsoneditor==0.1.4
protobuf==3.10.0
grpcio==1.25.0
protoc-gen-swagger==0.1.0
web3==4.8.1
m3u8==0.5.2
shortuuid==0.5.0
pylint-django==2.0.13